Updated SentimentEngine to work with paragraph. ANEW dictionary added
SentimentSentence was the old SentimeEngine. Now it incoporates all the
aggregation necessary to calculate it. Now takes in a wordlist parameter
to use which dictionary to use. Two dictionarys are included: anew and simple

Simple uses a negative word list and a postive word list to calculate sentiment.
anew uses a sentiment dictionary obtained from Eriq. See more in README.

CLI is now updated to work as a testing area. Need to incoporate error
checking done in CLI with SentimentEngine